Downloads/merra-two-master/requirements.txt
Flask            == 0.12.2
pandas           == 0.20.3
cassandra-driver == 3.10
mysql-connector  == 2.1.6
